requirements.txt
certifi==2018.10.15
chardet==3.0.4
commonmark==0.8.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
django-allauth==0.38.0
django-anymail==4.3
django-appconf==1.0.2
django-compressor==2.2
django-cors-headers==2.4.0
django-countries==5.3.2
django-crispy-forms==1.7.2
django-dnt==0.2.0
django-libsass==0.7
django-redis==4.9.0
django-wkhtmltopdf==3.2.0
django==2.1.2
djangorestframework==3.9.0
future==0.16.0
gunicorn==19.9.0
idna==2.7
itypes==1.1.0
jinja2==2.10
libsass==0.15.1
lml==0.0.4
markupsafe==1.0
mysqlclient==1.3.13
oauthlib==2.1.0
pyexcel-io==0.5.9.1
pyexcel==0.5.9.1
python3-openid==3.1.0
pytz==2018.5
rcssmin==1.0.6
redis==2.10.6
requests-oauthlib==1.0.0
requests==2.20.0
rjsmin==1.0.12
sentry-sdk==0.4.3
six==1.11.0
stripe==2.10.1
texttable==1.4.0
uritemplate==3.0.0
urllib3==1.24
whitenoise==4.1
